DAO fix

+++ src/main/java/by/siarhei/beerfest/dao/UserDAO.java
@@ -2,7 +2,7 @@ package by.siarhei.beerfest.dao;
 
 import by.siarhei.beerfest.entity.User;
 
-public interface UserDAO extends BaseDAO<Integer, User> {
+public interface UserDAO extends BaseDAO<Long, User> {
     User findUserByLogin(String login);
 
     boolean isExist(String login, String eMail);
